Update refactor project to use modules
