+++ src/modules/fromAPI.js
@@ -1,0 +1,14 @@
+import { baseURL, leaderBoardWrapper, displayLeaders } from "../index.js";
+
+const retrieveFromAPI = async () => {
+  const res = await fetch(baseURL);
+  const lead = await res.json();
+  const leaders = (await lead).result;
+
+  leaderBoardWrapper.innerHTML = '';
+  leaders.forEach((leader) => {
+    displayLeaders(leader);
+  });
+}
+
+export { retrieveFromAPI };
