--- src/modules/localstorage.js
@@ -1,12 +1,0 @@
-const addToLocalStorage = (myLeaderBoard) => {
-  localStorage.setItem('leaderBoard', JSON.stringify(myLeaderBoard));
-};
-
-const getFromLocalStorage = (myLeaderBoard) => {
-  if (localStorage.getItem('leaderBoard')) {
-    myLeaderBoard = JSON.parse(localStorage.getItem('leaderBoard'));
-  }
-  return myLeaderBoard;
-};
-
-export { addToLocalStorage, getFromLocalStorage };
